Added error messages
