+++ src/app/auth/signup/signup.component.ts
@@ -87,7 +87,7 @@ import { AuthService } from './../auth.service';
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators } from "@angular/forms";
 
-import { QuestionService } from '../../components/question/question.service';
+import { SignUpQuestionsService } from '../../components/questions/signUpQuestionsService.service';
 
 import { User } from '../../models/user'
 import { Router } from '@angular/router';
@@ -96,7 +96,7 @@ import { Router } from '@angular/router';
 	selector: 'app-signup',
 	templateUrl: './signup.component.html',
 	styleUrls: ['./signup.component.css'],
-	providers: [QuestionService]
+	providers: [SignUpQuestionsService]
 })
 
 export class SignupComponent {
@@ -107,7 +107,7 @@ export class SignupComponent {
 	constructor(
 		private authService: AuthService,
 		private router: Router,
-		service: QuestionService
+		service: SignUpQuestionsService
 	) {
 		this.questions = service.getQuestions();
 	}
@@ -127,3 +127,22 @@ export class SignupComponent {
 		});
 	}
 }
+
+
+// 		this.newUser.firstName = this.signUpForm.value.firstName;
+// 		this.newUser.lastName = this.signUpForm.get("lastName").value;
+// 		this.newUser.userName = this.signUpForm.get("userName").value;
+// 		this.newUser.street = this.signUpForm.get("street").value;
+// 		this.newUser.city = this.signUpForm.get("city").value;
+// 		this.newUser.postalCode = this.signUpForm.get("postalCode").value;
+// 		this.newUser.country = this.signUpForm.get("country").value;
+// 		this.newUser.mobilePhone = this.signUpForm.get("mobilePhone").value;
+// 		this.newUser.workPhone = this.signUpForm.get("workPhone").value;
+// 		this.newUser.homePhone = this.signUpForm.get("homePhone").value;
+// 		this.newUser.emergencyContactOnePhone = this.signUpForm.get("emergencyContactOnePhone").value;
+// 		this.newUser.emergencyContactTwoPhone = this.signUpForm.get("emergencyContactTwoPhone").value;
+// 		this.newUser.sailingQualifications = this.signUpForm.get("sailingQualifications").value;
+// 		this.newUser.skills = this.signUpForm.get("skills").value;
+// 		this.newUser.sailingExperience = this.signUpForm.get("sailingExperience").value;
+// 		this.newUser.email = this.signUpForm.get("email").value;
+// 		this.newUser.password = this.signUpForm.get("password").value;
