+++ src/app/components/dynamic-form-question/dynamic-form-question.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 
-import { QuestionBase } from '../question/question-base';
+import { QuestionBase } from '../questions/question-base';
 
 @Component({
 	selector: 'app-question',
@@ -9,7 +9,19 @@ import { QuestionBase } from '../question/question-base';
 	styleUrls: ['./dynamic-form-question.component.css'],
 })
 export class DynamicFormQuestionComponent {
+
 	@Input() question: QuestionBase<any>;
 	@Input() form: FormGroup;
-	get isValid() { return this.form.controls[this.question.key].valid; }
+
+	get isValid() {
+		return this.form.controls[this.question.key].valid;
+	}
+
+	get isDirty() {
+		return this.form.controls[this.question.key].dirty;
+	}
+
+	get isTouched() {
+		return this.form.controls[this.question.key].touched;
+	}
 }
