+++ src/app/components/dynamic-form/dynamic-form.component.html
@@ -9,10 +9,10 @@
 		<div>
 			<button class="btn btn-lg btn-primary btn-block" type="submit" [disabled]="form.invalid">Submit</button>
 		</div>
-	</form>
 
-	<div *ngIf="payLoad">
-		<strong>Saved the following values</strong><br>{{payLoad}}
-	</div>
+		<div *ngIf="payLoad">
+			<strong>Saved the following values</strong><br>{{payLoad}}
+		</div>
+	</form>
 
 </div>
