--- src/app/components/question/question.service.ts
@@ -1,34 +1,0 @@
-import { Injectable }       from '@angular/core';
-
-import { DropdownQuestion } from './question-dropdown';
-import { QuestionBase }     from './question-base';
-import { TextboxQuestion }  from './question-text';
-
-@Injectable()
-export class QuestionService {
-
-  // Todo: get from a remote source of question metadata
-  // Todo: make asynchronous
-  getQuestions() {
-
-    let questions: QuestionBase<any>[] = [
-
-      new TextboxQuestion({
-        key: 'firstName',
-        label: 'First name',
-        value: 'Bombasto',
-        required: true,
-        order: 1
-      }),
-
-      new TextboxQuestion({
-        key: 'emailAddress',
-        label: 'Email',
-        type: 'email',
-        order: 2
-      })
-    ];
-
-    return questions.sort((a, b) => a.order - b.order);
-  }
-}
