+++ src/app/components/questions/question-textbox.ts
@@ -1,7 +1,7 @@
 import { QuestionBase } from './question-base';
 
-export class TextboxQuestion extends QuestionBase<string> {
-  controlType = 'text';
+export class TextQuestion extends QuestionBase<string> {
+  controlType = 'textbox';
   type: string;
 
   constructor(options: {} = {}) {
